// src/app/obras/model/obra.ts
export class Obra {

    idobra: number;
    cei: string;
    nome: string;
    descricao: string;
    datainicio: Date;
    endereco: string;
    numero: string;
    bairro: string;
    cidade: string;
    cep: string;
    estado: string;
    engenheiro: string;
    previsaotermino: Date;
    datatermino: Date;
    padraoacabamento: string;
    blocos: string;
    unidades: number;
}